refactor: reesctructure repository to best fit common standards
