+++ tests/test_models/TestaPix.py
@@ -1,11 +1,6 @@
-import sys
-import os
-
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-
 import unittest
-from model.MetodoPagamento import MetodoPagamento
-from model.Pix import Pix
+from src.models.MetodoPagamento import MetodoPagamento
+from src.models.Pix import Pix
 
 class TestaPix(unittest.TestCase):
     """Classe que testa a classe Pix"""
